requirements.txt
Flask==2.2.2
google-generativeai==0.5.0
Werkzeug==2.2.2
